+++ spec/ig_markets/cli/config_file_spec.rb
@@ -1,30 +1,36 @@
 describe IGMarkets::CLI::ConfigFile do
   let(:config_file_contents) do
-    [
-      '--username=USERNAME',
-      '--password PASSWORD # The password',
-      '# A comment',
-      '--demo'
-    ]
+    {
+      'profiles' => {
+        'default' => { 'username' => 'A', 'password' => 'B', 'demo' => true },
+        'my-profile' => { 'username' => 'C', 'password' => 'D' }
+      }
+    }
   end
 
-  it 'prepends arguments to argv' do
-    config_file = IGMarkets::CLI::ConfigFile.new config_file_contents
+  let(:config_file) { IGMarkets::CLI::ConfigFile.new config_file_contents }
 
+  it 'prepends the default profile\'s arguments to argv' do
     argv = ['command', '--argument']
 
-    config_file.prepend_arguments_to_argv argv
+    config_file.prepend_profile_arguments_to_argv argv
 
-    expect(argv).to eq(%w(command --username=USERNAME --password PASSWORD --demo --argument))
+    expect(argv).to eq(%w(command --username=A --password=B --demo=true --argument))
   end
 
-  it 'finds the first config file that exists and opens it' do
+  it 'prepends a specified profile\'s arguments to argv' do
+    argv = ['command', '--argument', '--profile', 'my-profile']
+
+    config_file.prepend_profile_arguments_to_argv argv
+
+    expect(argv).to eq(%w(command --username=C --password=D --argument))
+  end
+
+  it 'finds the first config file that exists and loads it' do
     expect(File).to receive(:exist?).with('absent').and_return(false)
     expect(File).to receive(:exist?).with('present').and_return(true)
-    expect(File).to receive(:readlines).with('present').and_return(config_file_contents)
+    expect(YAML).to receive(:load_file).with('present').and_return(config_file_contents)
 
-    config_file = IGMarkets::CLI::ConfigFile.find 'absent', 'present', 'ignored'
-
-    expect(config_file.arguments).to eq(['--username=USERNAME', '--password', 'PASSWORD', '--demo'])
+    IGMarkets::CLI::ConfigFile.find 'absent', 'present', 'ignored'
   end
 end